One favourites menu on the player bar and full screen player
- drop the overflow menu button from the player bar
- always show the favourite and queue buttons on the desktop bar
- the heart opens a menu with add/remove favourite and add to playlist
- move the playback speed control to the full screen player's more options menu
