+++ src/layouts/default/PlayerOSD/Player.vue
@@ -18,22 +18,8 @@
       </div>
       <div class="mediacontrols-bottom-center">
         <div class="player-center-controls">
-          <div v-if="showWideCenterActions" class="player-center-side-action">
-            <Button
-              v-if="favoriteItem"
-              variant="ghost"
-              size="icon-lg"
-              class="player-control-button"
-              :aria-label="
-                $t(favoriteItem.favorite ? 'favorites_remove' : 'favorites_add')
-              "
-              @click="api.toggleFavorite(favoriteItem)"
-            >
-              <Heart
-                class="size-5"
-                :fill="favoriteItem.favorite ? 'currentColor' : 'none'"
-              />
-            </Button>
+          <div class="player-center-side-action">
+            <FavoriteMenuBtn class="player-control-button" />
           </div>
 
           <PlayerControls
@@ -53,7 +39,7 @@
             }"
           />
 
-          <div v-if="showWideCenterActions" class="player-center-side-action">
+          <div class="player-center-side-action">
             <QueueBtn :size="22" class="player-control-button" />
           </div>
         </div>
@@ -67,14 +53,6 @@
                the countdown to the full screen player and the player card's
                menu; it has no room of its own on the narrowest bars. -->
           <PlayerExtendedControls
-            :favorite="{
-              isVisible: false,
-              showInMenu: true,
-            }"
-            :queue="{
-              isVisible: false,
-              showInMenu: true,
-            }"
             :player="{
               isVisible: true,
             }"
@@ -107,13 +85,6 @@
       </div>
       <div class="mediacontrols-bottom-right">
         <div class="flex items-center">
-          <PlayerTrackMenu
-            v-if="showFloatingTrackMenu"
-            compact
-            force-visible
-            :show-favorite="true"
-            :show-queue="true"
-          />
           <!-- grouping sits beside play: both act on what this bar is playing -->
           <PlayerBarGroupControl floating />
           <!-- player mobile control buttons -->
@@ -156,18 +127,14 @@
 </template>
 
 <script setup lang="ts">
-import { Button } from "@/components/ui/button";
 import { ImageColorPalette, paletteFromServer } from "@/helpers/utils";
-import api from "@/plugins/api";
-import { MediaType } from "@/plugins/api/interfaces";
 import { getBreakpointValue } from "@/plugins/breakpoint";
 import { store } from "@/plugins/store";
 import vuetify from "@/plugins/vuetify";
-import { Heart } from "@lucide/vue";
 import { computed, ref, watch } from "vue";
 import PlayerBarGroupControl from "./PlayerBarGroupControl.vue";
 import PlayerBarMobileVolumeSheet from "./PlayerBarMobileVolumeSheet.vue";
-import PlayerTrackMenu from "./PlayerControlBtn/PlayerTrackMenu.vue";
+import FavoriteMenuBtn from "./PlayerControlBtn/FavoriteMenuBtn.vue";
 import QueueBtn from "./PlayerControlBtn/QueueBtn.vue";
 import PlayerControls from "./PlayerControls.vue";
 import PlayerExtendedControls from "./PlayerExtendedControls.vue";
@@ -180,15 +147,6 @@ interface Props {
 }
 const props = defineProps<Props>();
 const showMobileVolumeControls = ref(false);
-const showWideCenterActions = computed(() => getBreakpointValue("bp6"));
-const favoriteItem = computed(() => {
-  const item = store.curQueueItem?.media_item;
-  return item?.media_type === MediaType.AUDIO_SOURCE ? undefined : item;
-});
-
-// the floating row already carries grouping and play; the track menu only
-// joins them where there is width to spare
-const showFloatingTrackMenu = computed(() => getBreakpointValue("bp12"));
 
 /** Opens the fullscreen player, or the player picker when there is nothing to show. */
 function openActivePlayer() {
@@ -448,37 +406,6 @@ watch(
   }
 }
 
-/* the sleep timer joins the actions as a fifth control, which the widths above
-   do not budget for: they need a bar of about 1221px before the five of them
-   fit, and until then the timer reaches over the timeline. So while one runs
-   the actions take the widths they use below 1100px, which fit from 1034px up.
-   The full widths take over again at 1250px, where they clear the timeline by
-   about 9px - the tightest the row gets, so a sixth control or a wider button
-   would have to raise this bound with it. */
-@media screen and (min-width: 1100px) and (max-width: 1249px) {
-  .mediacontrols :deep(.player-bar-action-row:has(.player-bar-sleep-timer)) {
-    .player-bar-menu-button {
-      width: 40px !important;
-    }
-
-    .player-bar-volume-button {
-      width: 56px !important;
-    }
-
-    /* the floor these two carry above 1100px would outrank the width, so it
-       has to be reset for the width to take effect */
-    .player-bar-group-button {
-      width: 60px !important;
-      min-width: 0 !important;
-    }
-
-    .player-bar-player-button {
-      width: 68px !important;
-      min-width: 0 !important;
-    }
-  }
-}
-
 @media screen and (min-width: 769px) and (max-width: 1099px) {
   .mediacontrols .mediacontrols-bottom-center {
     width: auto;
@@ -486,11 +413,6 @@ watch(
 
   .mediacontrols-bottom-right {
     min-width: 0;
-  }
-
-  .mediacontrols :deep(.player-bar-menu-button) {
-    width: 40px !important;
-    height: 72px !important;
   }
 
   .mediacontrols :deep(.player-bar-volume-button) {
