+++ src/layouts/default/PlayerOSD/PlayerFullscreen.vue
@@ -367,19 +367,12 @@
 
         <!-- main media control buttons (play, next, previous etc.)-->
         <div class="media-controls">
-          <Icon
-            :disabled="!store.curQueueItem?.media_item"
-            :title="$t('tooltip.favorite')"
-            variant="button"
-            class="media-controls-item"
-            max-height="30px"
-            @click="onHeartBtnClick"
-          >
-            <Heart
+          <div class="media-controls-item favorite-btn-wrapper">
+            <FavoriteMenuBtn
+              style="max-height: 30px; min-height: 0; min-width: 0"
               :size="18"
-              :fill="currentItemFavorite ? 'currentColor' : 'none'"
             />
-          </Icon>
+          </div>
           <ShuffleBtn
             v-if="$vuetify.display.mdAndUp"
             :player-queue="store.activePlayerQueue"
@@ -475,12 +468,13 @@
           </Button>
         </div>
       </div>
+
+      <PlaybackSpeedDialog v-model:open="playbackSpeedDialogOpen" />
     </v-card>
   </v-dialog>
 </template>
 
 <script setup lang="ts">
-import Icon from "@/components/Icon.vue";
 import LyricsViewer from "@/components/LyricsViewer.vue";
 import MarqueeText from "@/components/MarqueeText.vue";
 import PanelDragHandle from "@/components/PanelDragHandle.vue";
@@ -493,7 +487,6 @@ import { useActiveTrackWaveform } from "@/composables/useActiveTrackWaveform";
 import { setStatusBarColorOverride } from "@/composables/useStatusBarColor";
 import { useUserPreferences } from "@/composables/userPreferences";
 import { useVisualizer } from "@/composables/visualizer/useVisualizer";
-import type { ContextMenuItem } from "@/helpers/context_menu_item";
 import { MarqueeTextSync } from "@/helpers/marquee_text_sync";
 import { getPlayerMenuItems } from "@/helpers/player_menu_items";
 import {
@@ -503,6 +496,8 @@ import {
   getPlayerName,
 } from "@/helpers/utils";
 import LyricsOffsetMenuControl from "@/layouts/default/PlayerOSD/LyricsOffsetMenuControl.vue";
+import PlaybackSpeedDialog from "@/layouts/default/PlayerOSD/PlaybackSpeedDialog.vue";
+import FavoriteMenuBtn from "@/layouts/default/PlayerOSD/PlayerControlBtn/FavoriteMenuBtn.vue";
 import NextBtn from "@/layouts/default/PlayerOSD/PlayerControlBtn/NextBtn.vue";
 import PlayBtn from "@/layouts/default/PlayerOSD/PlayerControlBtn/PlayBtn.vue";
 import PreviousBtn from "@/layouts/default/PlayerOSD/PlayerControlBtn/PreviousBtn.vue";
@@ -516,21 +511,14 @@ import VisualizerMenuControl from "@/layouts/default/PlayerOSD/VisualizerMenuCon
 import { useFullscreenQueue } from "@/layouts/default/PlayerOSD/useFullscreenQueue";
 import api from "@/plugins/api";
 import { getSourceName } from "@/plugins/api/helpers";
-import {
-  EventMessage,
-  EventType,
-  MediaItemType,
-  MediaType,
-  PlayerType,
-  Track,
-} from "@/plugins/api/interfaces";
+import { MediaType, PlayerType, Track } from "@/plugins/api/interfaces";
 import { getBreakpointValue } from "@/plugins/breakpoint";
 import { eventbus } from "@/plugins/eventbus";
 import { $t } from "@/plugins/i18n";
 import router from "@/plugins/router";
 import { store } from "@/plugins/store";
 import vuetify from "@/plugins/vuetify";
-import { ChevronDownIcon, EllipsisVerticalIcon, Heart } from "@lucide/vue";
+import { ChevronDownIcon, EllipsisVerticalIcon } from "@lucide/vue";
 import Color from "color";
 import {
   computed,
@@ -596,19 +584,6 @@ watch(
     el.style.removeProperty("transition");
     el.style.removeProperty("will-change");
   },
-);
-
-// Track the favorite state of the current queue item independently from
-// media_item.favorite so optimistic updates survive server-side queue refreshes
-// (which replace the whole current_item object and would reset the field).
-const currentItemFavorite = ref(false);
-watch(
-  () => store.curQueueItem?.queue_item_id,
-  () => {
-    currentItemFavorite.value =
-      store.curQueueItem?.media_item?.favorite ?? false;
-  },
-  { immediate: true },
 );
 
 // Local reactive state for lyrics
@@ -1158,6 +1133,16 @@ const onArtistClick = async function () {
   }
 };
 
+const playbackSpeedDialogOpen = ref(false);
+
+// only audiobooks and podcast episodes carry a per-item playback speed
+const playbackSpeedSupported = computed(() => {
+  const mediaType = store.curQueueItem?.media_item?.media_type;
+  return (
+    mediaType === MediaType.AUDIOBOOK || mediaType === MediaType.PODCAST_EPISODE
+  );
+});
+
 const openQueueMenu = function (evt: Event) {
   if (!store.activePlayer) return;
   const menuItems = getPlayerMenuItems(
@@ -1168,6 +1153,17 @@ const openQueueMenu = function (evt: Event) {
       hideShuffleRepeat: mdAndUp.value,
     },
   );
+  // playback speed only means something for spoken-word content, which is also
+  // the only content anyone plays at anything but 1x
+  if (playbackSpeedSupported.value) {
+    menuItems.push({
+      label: "change_playback_speed",
+      action: () => {
+        playbackSpeedDialogOpen.value = true;
+      },
+      icon: "mdi-speedometer",
+    });
+  }
   // The waveform toggle slots in above the visualizer entries, keeping the
   // visualizer toggle + options grouped at the bottom of the menu. The on/off
   // toggle itself comes from getPlayerMenuItems (it is a player control,
@@ -1213,23 +1209,6 @@ const openQueueMenu = function (evt: Event) {
   });
 };
 
-// sync currentItemFavorite when the server confirms a favorite change
-onMounted(() => {
-  const unsub = api.subscribe(
-    EventType.MEDIA_ITEM_UPDATED,
-    (evt: EventMessage) => {
-      const updatedItem = evt.data as MediaItemType;
-      if (
-        "favorite" in updatedItem &&
-        store.curQueueItem?.media_item?.uri === updatedItem.uri
-      ) {
-        currentItemFavorite.value = updatedItem.favorite;
-      }
-    },
-  );
-  onBeforeUnmount(unsub);
-});
-
 // Handle Escape key to close fullscreen player (since persistent disables default behavior)
 const onKeydown = (e: KeyboardEvent) => {
   if (
@@ -1247,103 +1226,6 @@ onMounted(() => {
     window.removeEventListener("keydown", onKeydown);
   });
 });
-
-const onHeartBtnClick = async function (evt: PointerEvent | MouseEvent) {
-  // the heart icon/button was clicked
-  if (!store.curQueueItem?.media_item) return;
-  if (!currentItemFavorite.value) {
-    currentItemFavorite.value = true; // optimistic — survives server queue refresh
-    api.toggleFavorite(store.curQueueItem.media_item);
-    return;
-  }
-  //resolve media item into a library item
-  let resolvedItem = store.curQueueItem!.media_item as MediaItemType;
-  if (
-    (resolvedItem.provider != "library" ||
-      !("provider_mappings" in resolvedItem)) &&
-    [
-      MediaType.ALBUM,
-      MediaType.ARTIST,
-      MediaType.AUDIOBOOK,
-      MediaType.PLAYLIST,
-      MediaType.PODCAST,
-      MediaType.RADIO,
-      MediaType.TRACK,
-    ].includes(resolvedItem.media_type)
-  ) {
-    // resolve itemmapping or non-library item
-    resolvedItem =
-      (await api.getLibraryItem(
-        resolvedItem.media_type,
-        resolvedItem.item_id,
-        resolvedItem.provider,
-      )) || resolvedItem;
-  }
-
-  const menuItems: ContextMenuItem[] = [];
-
-  // remove from favorites
-  if (
-    "favorite" in resolvedItem &&
-    resolvedItem.favorite &&
-    resolvedItem.provider == "library" &&
-    [
-      MediaType.ALBUM,
-      MediaType.ARTIST,
-      MediaType.AUDIOBOOK,
-      MediaType.PLAYLIST,
-      MediaType.PODCAST,
-      MediaType.RADIO,
-      MediaType.TRACK,
-    ].includes(resolvedItem.media_type)
-  ) {
-    menuItems.push({
-      label: "favorites_remove",
-      labelArgs: [],
-      action: () => {
-        api.removeItemFromFavorites(
-          resolvedItem.media_type,
-          resolvedItem.item_id,
-        );
-        resolvedItem.favorite = false;
-        store.curQueueItem!.media_item!.favorite = false;
-        currentItemFavorite.value = false;
-      },
-      icon: "mdi-heart",
-    });
-  }
-
-  // add to playlist
-  if (
-    [
-      MediaType.ALBUM,
-      MediaType.ARTIST,
-      MediaType.AUDIOBOOK,
-      MediaType.PLAYLIST,
-      MediaType.PODCAST,
-      MediaType.RADIO,
-      MediaType.TRACK,
-    ].includes(resolvedItem.media_type)
-  ) {
-    menuItems.push({
-      label: "add_playlist",
-      labelArgs: [],
-      action: () => {
-        eventbus.emit("playlistdialog", {
-          items: [store.curQueueItem!.media_item as MediaItemType],
-        });
-      },
-      icon: "mdi-plus-circle-outline",
-    });
-  }
-
-  // open the contextmenu by emitting the event
-  eventbus.emit("contextmenu", {
-    items: menuItems,
-    posX: evt.clientX,
-    posY: evt.clientY,
-  });
-};
 
 const sliderColor = ref<string | undefined>(undefined);
 const backgroundColor = ref<string | undefined>(undefined);
@@ -1717,7 +1599,8 @@ onBeforeUnmount(() => {
   height: 100%;
 }
 
-.queue-btn-wrapper {
+.queue-btn-wrapper,
+.favorite-btn-wrapper {
   display: flex;
   align-items: center;
   justify-content: center;
@@ -1763,7 +1646,8 @@ onBeforeUnmount(() => {
 }
 
 /* Match the marginless transport icons so the play button stays centred. */
-.media-controls > .queue-btn-wrapper {
+.media-controls > .queue-btn-wrapper,
+.media-controls > .favorite-btn-wrapper {
   margin: 0;
 }
