+++ tests/layouts/Player.test.ts
@@ -3,7 +3,8 @@ import PlayerBarGroupControl from "@/layouts/default/PlayerOSD/PlayerBarGroupCon
 import PlayerBarMobileVolumeSheet from "@/layouts/default/PlayerOSD/PlayerBarMobileVolumeSheet.vue";
 import PlayerExtendedControls from "@/layouts/default/PlayerOSD/PlayerExtendedControls.vue";
 import PlayerTrackDetails from "@/layouts/default/PlayerOSD/PlayerTrackDetails.vue";
-import PlayerTrackMenu from "@/layouts/default/PlayerOSD/PlayerControlBtn/PlayerTrackMenu.vue";
+import FavoriteMenuBtn from "@/layouts/default/PlayerOSD/PlayerControlBtn/FavoriteMenuBtn.vue";
+import QueueBtn from "@/layouts/default/PlayerOSD/PlayerControlBtn/QueueBtn.vue";
 import PlayerVolume from "@/layouts/default/PlayerOSD/PlayerVolume.vue";
 import type { Player as PlayerModel } from "@/plugins/api/interfaces";
 import { store } from "@/plugins/store";
@@ -41,11 +42,12 @@ const mockStore = store as unknown as {
   activePlayerId?: string;
 };
 
-// bp12: the width from which the floating row has room for the track menu
-const TRACK_MENU_BREAKPOINT = 415;
-
 // bp7: the width from which the desktop action row has room for the sleep timer
 const SLEEP_TIMER_BREAKPOINT = 1100;
+
+// the width below which the floating bar takes over from the desktop one, so
+// the narrowest the desktop bar is ever laid out at
+const NARROWEST_DESKTOP_BAR = 769;
 
 const originalInnerWidth = Object.getOwnPropertyDescriptor(
   window,
@@ -110,20 +112,6 @@ describe("Player floating mobile bar", () => {
     );
   });
 
-  it.each([
-    { width: TRACK_MENU_BREAKPOINT, present: true },
-    { width: TRACK_MENU_BREAKPOINT - 1, present: false },
-  ])("shows the track menu at $width px: $present", ({ width, present }) => {
-    mockStore.activePlayer = { player_id: "p1" } as PlayerModel;
-    setViewportWidth(width);
-    const player = mountPlayer(true);
-
-    expect(player.findComponent(PlayerTrackMenu).exists()).toBe(present);
-    // the menu is the only control the narrow row gives up
-    expect(player.findComponent(PlayerBarGroupControl).exists()).toBe(true);
-    expect(player.findComponent(PlayerVolume).exists()).toBe(true);
-  });
-
   it("closes an open volume sheet when the active player changes", async () => {
     mockStore.activePlayer = { player_id: "p1" } as PlayerModel;
     mockStore.activePlayerId = "p1";
@@ -168,4 +156,18 @@ describe("Player desktop bar", () => {
       player.findComponent(PlayerExtendedControls).props("sleepTimer"),
     ).toEqual({ isVisible: visible });
   });
+
+  // the favourite and queue buttons flank the transport controls at every width
+  // the desktop bar is laid out at - narrower than this it is not the bar in use
+  it.each([NARROWEST_DESKTOP_BAR, 1500])(
+    "keeps the favourite and queue buttons at %i px",
+    (width) => {
+      mockStore.activePlayer = { player_id: "p1" } as PlayerModel;
+      setViewportWidth(width);
+      const player = mountPlayer(false);
+
+      expect(player.findComponent(FavoriteMenuBtn).exists()).toBe(true);
+      expect(player.findComponent(QueueBtn).exists()).toBe(true);
+    },
+  );
 });
